Move model logic to view
